Rewrite sim to use multiple threads
